+++ _freeze/analysis/visualizing-risk/execute-results/html.json
@@ -2,10 +2,8 @@
   "hash": "768b99dcc830c2005991e9caf67d5fbd",
   "result": {
     "engine": "knitr",
-    "markdown": "---\ntitle: \"Visualizing Risk (Draft)\"\nauthor: \"John Benninghoff\"\ndate: '2024-12-06'\ndate-modified: '2025-02-25'\ncategories: risk quantification\norder: 106\noutput:\n  html_notebook:\n    theme:\n      version: 5\n      preset: bootstrap\n    css: assets/extra.css\n    pandoc_args: --shift-heading-level-by=1\n    toc: yes\n    toc_float:\n      collapsed: no\n      smooth_scroll: no\n---\n\n\n\n\nAn exploration of different approaches to communicating security breach risk informed by prior work.\n\n# Questions/TODO\n\n- [ ] How should tail risk be visualized?\n- [ ] How well do quantile dotplots capture tail risk?\n- [ ] Review Open Group work on cyber risk\n  - [ ] [Calculating Reserves for Cyber Risk: Integrating Cyber Risk with Financial Risk](https://riskgenesis.com/wp-content/uploads/2022/02/OG-Doc-w215.pdf)\n  - [ ] [Calculating Reserves for Cyber Risk: An Open FAIR Approach](https://www.opengroup.org/sites/default/files/pdf/Cyber_Risk_Capital_Requirements_Project_Charter_-_V1.0.pdf)\n\n\n\n\n::: {.cell}\n\n```{.r .cell-code}\n# no libraries\n```\n:::\n\n\n\n\n# Background\n\nThe [Risk Value Analysis](risk-value.Rmd) found that a firm's executive is unlikely to invest in\nreducing the risk of a large security breach, suggesting that security spending should be presented\ndifferently:\n\n> At the firm level, I think this means that security leaders shouldn't present security as an\ninvestment. As with safety, I think the main argument for better security is a moral or emotional\ncase: we care about security because we care about our customers, partners, and other stakeholders.\nAlso, people are typically loss-averse, so expressing security risk in those terms will better\nconnect with decision makers. Using\n[Tail value at risk](https://en.wikipedia.org/wiki/Tail_value_at_risk) or\n[Loss Exceedance Curves](https://www.cyentia.com/communicating-risk-loss-exceedance-curves/) express\nloss in this way - \"There's a 5% chance of cybersecurity losses exceeding \\$780,000 and a 1% chance\nof losses exceeding \\$25,000,000 over the next year.\" I also think it means security leaders should\nbe mindful of *how* they spend their limited funds, by maximizing investments in\n[what works](https://doi.org/10.1080/23738871.2024.2335461).\n\nThis analysis explores different approaches to communicating that risk with the goal of promoting\ndecisions on security spending so that all participants are satisfied with the outcome, informed by\nprior work on risk communication and risk perception, including the work of\n[Lace Padilla](http://www.lacepadilla.com), and Cyentia's\n[analysis](https://www.youtube.com/watch?v=eiMBMsN_4j4) on tail value at risk.\n",
-    "supporting": [
-      "visualizing-risk_files"
-    ],
+    "markdown": "---\ntitle: \"Visualizing Risk (Draft)\"\nauthor: \"John Benninghoff\"\ndate: '2024-12-06'\ndate-modified: '2025-02-25'\ncategories: risk quantification\norder: 106\noutput:\n  html_notebook:\n    theme:\n      version: 5\n      preset: bootstrap\n    css: assets/extra.css\n    pandoc_args: --shift-heading-level-by=1\n    toc: yes\n    toc_float:\n      collapsed: no\n      smooth_scroll: no\n---\n\nAn exploration of different approaches to communicating security breach risk informed by prior work.\n\n# Questions/TODO\n\n- [ ] How should tail risk be visualized?\n- [ ] How well do quantile dotplots capture tail risk?\n- [ ] Review Open Group work on cyber risk\n  - [ ] [Calculating Reserves for Cyber Risk: Integrating Cyber Risk with Financial Risk](https://riskgenesis.com/wp-content/uploads/2022/02/OG-Doc-w215.pdf)\n  - [ ] [Calculating Reserves for Cyber Risk: An Open FAIR Approach](https://www.opengroup.org/sites/default/files/pdf/Cyber_Risk_Capital_Requirements_Project_Charter_-_V1.0.pdf)\n\n\n::: {.cell}\n\n```{.r .cell-code}\n# no libraries\n```\n:::\n\n\n# Background\n\nThe [Risk Value Analysis](risk-value.Rmd) found that a firm's executive is unlikely to invest in\nreducing the risk of a large security breach, suggesting that security spending should be presented\ndifferently:\n\n> At the firm level, I think this means that security leaders shouldn't present security as an\ninvestment. As with safety, I think the main argument for better security is a moral or emotional\ncase: we care about security because we care about our customers, partners, and other stakeholders.\nAlso, people are typically loss-averse, so expressing security risk in those terms will better\nconnect with decision makers. Using\n[Tail value at risk](https://en.wikipedia.org/wiki/Tail_value_at_risk) or\n[Loss Exceedance Curves](https://www.cyentia.com/communicating-risk-loss-exceedance-curves/) express\nloss in this way - \"There's a 5% chance of cybersecurity losses exceeding \\$780,000 and a 1% chance\nof losses exceeding \\$25,000,000 over the next year.\" I also think it means security leaders should\nbe mindful of *how* they spend their limited funds, by maximizing investments in\n[what works](https://doi.org/10.1080/23738871.2024.2335461).\n\nThis analysis explores different approaches to communicating that risk with the goal of promoting\ndecisions on security spending so that all participants are satisfied with the outcome, informed by\nprior work on risk communication and risk perception, including the work of\n[Lace Padilla](http://www.lacepadilla.com), and Cyentia's\n[analysis](https://www.youtube.com/watch?v=eiMBMsN_4j4) on tail value at risk.\n",
+    "supporting": [],
     "filters": [
       "rmarkdown/pagebreak.lua"
     ],
