+++ docs/sitemap.xml
@@ -18,7 +18,7 @@
   </url>
   <url>
     <loc>https://jabenninghoff.github.io/security/analysis/rq-demo.html</loc>
-    <lastmod>2025-04-18T20:58:24.295Z</lastmod>
+    <lastmod>2025-08-13T16:12:54.808Z</lastmod>
   </url>
   <url>
     <loc>https://jabenninghoff.github.io/security/analysis/breach-rates.html</loc>
@@ -42,7 +42,7 @@
   </url>
   <url>
     <loc>https://jabenninghoff.github.io/security/NEWS.html</loc>
-    <lastmod>2025-04-18T20:55:55.410Z</lastmod>
+    <lastmod>2025-08-13T16:19:52.238Z</lastmod>
   </url>
   <url>
     <loc>https://jabenninghoff.github.io/security/index.html</loc>
